+++ ketodiet_django_project/user/models.py
@@ -1,7 +1,7 @@
 from django.db import models
 
 class UserDB(models.Model):
-    id = models.IntegerField(primary_key=True)
+    id = models.BigIntegerField(primary_key=True)
     name = models.CharField(max_length=12)
     isAdmin = models.BooleanField(default=False)
